1st upgrade version
1st upgrade version

+++ MagicQuiz/AlexTools/Ads/ChartBoost/JSON3/SBJson3StreamParser.m
@@ -119,7 +119,7 @@
 }
 
 - (void)maxDepthError {
-    self.error = [NSString stringWithFormat:@"Input depth exceeds max depth of %lu", maxDepth];
+    self.error = [NSString stringWithFormat:@"Input depth exceeds max depth of %lu", (unsigned long)maxDepth];
     self.state = [SBJson3StreamParserStateError sharedInstance];
 }
